Rename files to better match their exported classes
