+++ lib/DetailedError.ts
