+++ lib/browser/XHRHttpStack.ts
