+++ lib/logger.ts
@@ -1,5 +1,6 @@
 let isEnabled = false
 
+// TODO: Replace this global state with an option for the Upload class
 export function enableDebugLog(): void {
   isEnabled = true
 }
