+++ lib/node/NodeFileReader.ts
